+++ docs/api.rst
@@ -90,9 +90,6 @@ Datasets
     BaseConcatDataset
     WindowsDataset
     MOABBDataset
-    create_from_X_y
-    create_from_mne_raw
-    create_from_mne_epochs
 
 
 Data Utils
@@ -105,6 +102,9 @@ Data Utils
 .. autosummary::
    :toctree: generated/
 
+    create_from_X_y
+    create_from_mne_raw
+    create_from_mne_epochs
     create_fixed_length_windows
     create_windows_from_events
     exponential_moving_demean
