+++ environment.yml
@@ -14,7 +14,7 @@ dependencies:
 - pytorch
 - pytest
 - pip
-- skorch
+- skorch==0.7
 - mne
 - pip:
   - https://github.com/braindecode/braindecode/zipball/master
