+++ examples/plot_mne_dataset_example.py
@@ -11,7 +11,7 @@
 
 import mne
 
-from braindecode.datasets.mne import (
+from braindecode.datautil.mne import (
     create_from_mne_raw, create_from_mne_epochs)
 
 ###############################################################################
