+++ requirements.txt
@@ -4,4 +4,4 @@ scipy
 matplotlib
 h5py
 mne
-skorch
+skorch==0.7
