+++ test/unit_tests/datasets/test_mne.py
@@ -1,7 +1,7 @@
 import numpy as np
 import mne
 
-from braindecode.datasets.mne import  (
+from braindecode.datautil.mne import  (
     create_from_mne_raw, create_from_mne_epochs)
 
 def test_create_from_single_raw():
